Sound buffer generation code implemented but not tested

+++ SynthExperiments/jni/Envelope.cpp
@@ -16,7 +16,7 @@ Envelope::Envelope()
 	releaseRamp = NULL;
 }
 
-Envelope::Envelope( float sustainLevel )
+Envelope::Envelope( double sustainLevel )
 {
 	sustain = sustainLevel;
 	releaseRamp = NULL;
